initial commit - all works
this application is intended for use as an initilization container
for a kubernetes pod that will be streaming application logs to a
central service.  the central service can be found in k2-charts

the shell script expect two volumes to be mounted:
 - one is an emptyDir that is pod unique and will be shared between
the initialization pod and the application pod.  it will contain a
single symlink (application) that will point to the other volume.
this is where applications should write their logs to
 - two is a hostPath directory that is shared between both all
application pods that want to centralize their logs and the log
aggregation pod.  this will contain many directories, one for each
application pod.

the idea is that all applications will write to the common link in
the emptyDir volume.  the emptyDir is unique to each pod so no
collisions.  the hostPath mount will have a directory, created by
this script, that is unique to each pod (pod + namespace are in the
directory name).  The log aggregator will then use the directory
name to attach the pod name and namespace to each outgoing log event.